fix: PostScript integers are 64-bit, matching Ghostscript
The standard LCG idiom PostScript programs use for pseudo-randomness,

    /seed seed 1103515245 mul 12345 add 2147483648 mod def

failed on a 32-bit integer. 20260711 * 1103515245 overflows, promotes to
a real per PLRM, and `mod` — which is integer-only — then raises
typecheck. Ghostscript renders these files; stet refused them.

Widening only the real fallback would not have fixed it, and that is
what settles the design. The exact product is 22358003463039195, which
needs 55 bits; f64 carries 53 and rounds it to ...196. The seed comes
out 1447897308 where Ghostscript and exact arithmetic give 1447897307,
so the file would have rendered a different tree on the first draw with
no error to say so. Correctness here requires the integer, not a wider
approximation of it.

PLRM Appendix B's 32-bit range is not a conformance requirement. It is
listed under "Typical Limits" as an architectural limit of interpreters
"running on 32-bit machines" which "do not necessarily apply to all
PostScript implementations". Ghostscript is 64-bit — `2147483647 1 add`
is `2147483648`, integertype — and PostForge, the reference
implementation, uses Python ints and is effectively unbounded. stet was
the outlier.

Costs nothing in memory: `Real(f64)` already forces an 8-byte payload
and 8-byte alignment, so `PsValue` and `PsObject` are the same size as
before. `PsObject::int` takes `impl Into<i64>`, which left the ~550
existing call sites untouched. `as_i32()` now range-checks and returns
None rather than truncating — its callers are array and string indices
and character codes, all genuinely bounded, and a value too large to be
one should fail their range check rather than wrap into a valid-looking
index. `as_i64()` is added for the full range. Operators that take
bounded parameters (setlinecap, setlinejoin, setobjectformat,
SubFileDecode's count) narrow explicitly and raise rangecheck.

`bitshift` widens to 64 bits with it, so `1 40 bitshift` is 2^40 rather
than 0, matching Ghostscript. Right shift stays logical per PLRM ("bits
shifted in are 0"); the same paragraph notes the operation "produces an
arithmetically correct result only for positive values of int1", so the
negative case is explicitly undefined and left as it was.

A differential probe against Ghostscript across the integer-touching
operators then caught a second bug: `cvi` on a long integer-valued
string was off by one, because the string scanner returned f64 and
`(22358003463039195) cvi` did not survive the round trip. Integer
literals now stay integral.

Verified against Ghostscript 10.05.1: 14 expected values in the new PS
tests match exactly, and the probe is now identical on every integer
operator. A sample corpus of 60 generative PostScript programs renders
clean where two previously failed, and a recursive-tree program that
seeds this LCG produces the same tree as gs — same branch topology, same
blossom placement — which is the check that the seed sequence agrees.

The remaining probe differences are real-number formatting: gs prints
large reals in scientific notation and uses single-precision, so it
renders 123456789.0 as 123456792.0 where stet prints it exactly. That
divergence predates this change and is untouched.

+++ crates/stet-core/src/context.rs
@@ -140,7 +140,9 @@ pub struct Context {
 
     // State
     pub rand_state: u64,
-    pub rand_seed: i32,
+    /// Last value handed to `srand`, returned by `rrand`. Widened with
+    /// [`PsValue::Int`]: `rrand` must return exactly what `srand` was given.
+    pub rand_seed: i64,
     /// Current source line number (1-based), updated during scanning.
     pub current_source_line: u32,
     /// Packing mode for array/procedure creation (setpacking/currentpacking).
