+++ crates/stet-core/src/system_font_loader.rs
@@ -419,7 +419,7 @@ fn register_cff_font(
         }
 
         if cff_font.is_cid {
-            ctx.dicts.put(cs_entity, DictKey::Int(gid as i32), cs_obj);
+            ctx.dicts.put(cs_entity, DictKey::Int(gid as i64), cs_obj);
         }
     }
     ctx.dicts.put(
